+++ src/constants.ts
@@ -1,4 +1,4 @@
-export const DOCTOR_VERSION = '0.2.0'
+export const DOCTOR_VERSION = '0.2.1'
 export const SUPPORTED_DSH_VERSION = '0.1.0-rc.6'
 export const PLAN_TTL_MS = 10 * 60 * 1000
 export const CHECKPOINT_RETENTION = 10